+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 
 - fix: Rate limiting for cached envelope items #685
+- feat: Errors and sessions in the same envelope #686
+- feat: Implement NSCopying for SentrySession #683
 - fix: Crash when SentryClient is nil in SentryHub #681
 - feat: Send cached envelopes first #676
 - feat: Use envelopes for sending events #650